Keep source event statement counts in their native index type

+++ crates/fe2o3-lower-mir-kernel/src/production_execution_lifecycle_producer_v29_tests.rs
@@ -419,7 +419,7 @@ fn run_lifecycle(
             (
                 HELPER,
                 0,
-                semantic.functions()[1].blocks()[0].statements().len() as u32,
+                semantic.functions()[1].blocks()[0].statements().len(),
                 ProductionScopeEventKindV29::Call {
                     callee: derive.callee(),
                     kind: ProductionScopeCallKindV29::Derive,
